requirements.txt
pymongo==4.6.1
python-dotenv==1.0.0
tqdm==4.66.1
humanize==4.9.0
questionary==2.0.1
rich==13.7.0
pytest
testcontainers[mongodb]>=3.7.1